+++ .idea/bfb_vote_counter.iml
@@ -6,5 +6,6 @@
     </content>
     <orderEntry type="jdk" jdkName="Python 3.8" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="waterbot" />
   </component>
 </module>
