+++ config.json
@@ -4,31 +4,31 @@
     "token": "AIzaSyDD2Y-CM7yKTTLJb7WyTc2NJR2djv28TY4",
     "characters": {
         "a": "9 ball",
-        "b": "anchor",
-        "c": "avocado ",
-        "d": "battery",
-        "e": "blender",
-        "f": "boom mic",
-        "g": "camera",
-        "h": "clapboard",
-        "i": "conch shell",
-        "j": "discy",
-        "k": "income tax return document",
-        "l": "kitchen sink",
-        "m": "leek",
-        "n": "nonexisty ",
-        "o": "onigiri",
-        "p": "pda",
-        "q": "price tag",
-        "r": "rubber spatula",
-        "s": "salt lamp",
-        "t": "scissors",
-        "u": "shampoo",
-        "v": "shooping cart",
-        "w": "snare sum",
-        "x": "tape",
+        "b": "Anchor",
+        "c": "Avocado ",
+        "d": "Battery",
+        "e": "Blender",
+        "f": "Boom mic",
+        "g": "Camera",
+        "h": "Chapboard",
+        "i": "Conch Shell",
+        "j": "Discy",
+        "k": "ICRD",
+        "l": "Kitchen sink",
+        "m": "Leek",
+        "n": "Nonexisty ",
+        "o": "Onigiri",
+        "p": "PDA",
+        "q": "Price Tag",
+        "r": "Rubber Spatula",
+        "s": "Salt Lamp",
+        "t": "Scissors",
+        "u": "Shampoo",
+        "v": "Shopping Cary",
+        "w": "Snare Drum",
+        "x": "Tape",
         "y": "VHSy",
-        "Z": "Winner"
+        "z": "Winner"
     },
     "deadline": 172800
 }
