+++ config.json~
@@ -1,0 +1,34 @@
+{
+    "info": "bfb 16 rc debut config",
+    "videoID": "gs3lqEJHSnA",
+    "token": "bbbbbAIzaSyDD2Y-CM7yKTTLJb7WyTc2NJR2djv28TY4",
+    "characters": {
+        "a": "9 ball",
+        "b": "Anchor",
+        "c": "Avocado ",
+        "d": "Battery",
+        "e": "Blender",
+        "f": "Boom mic",
+        "g": "Camera",
+        "h": "Chapboard",
+        "i": "Conch Shell",
+        "j": "Discy",
+        "k": "ICRD",
+        "l": "Kitchen sink",
+        "m": "Leek",
+        "n": "Nonexisty ",
+        "o": "Onigiri",
+        "p": "PDA",
+        "q": "Price Tag",
+        "r": "Rubber Spatula",
+        "s": "Salt Lamp",
+        "t": "Scissors",
+        "u": "Shampoo",
+        "v": "Shopping Cary",
+        "w": "Snare Drum",
+        "x": "Tape",
+        "y": "VHSy",
+        "z": "Winner"
+    },
+    "deadline": 172800
+}
