+++ counter.py
@@ -1,399 +1,226 @@
-# Main file
-import argparse
-import json
 import os
-import pickle
 import random
 import re
+import json
 import sys
-import traceback
-from time import sleep
-from time import time
-from datetime import datetime
-
+import time
+import logging
 import requests
-from colorama import Cursor
-from colorama import Fore
-from colorama import init
-from colorama import Style
+import argparse
+from colorama import Fore, Style, init
 
 import gen
 
-# DEBUG
+init_time = time.time()
+version = ('0', '1-alpha', 'rewrite-1')
+ver = "%s.%s.%s" % version
 
-version = ("0.", "1-alpha.", "3")
+# Parse arguments
 init(autoreset=True)
+print(f'{Fore.CYAN}BFB Vote Counter {Fore.YELLOW}v%s\n' % ver)
+parser = argparse.ArgumentParser(description="Simple python script for counting votes in Battle for Dream Island "
+                                             "season 4 and (potentially) 5, a popular animated object show on YouTube.")
+parser.add_argument('-v', '--version',
+                    help="Prints the version and exits", action='version', version='\033[2A')
+parser.add_argument('-f', '--comments-file',
+                    help="The comment pickle file when it finished getting the votes and/or "
+                         "it errored out. It will look something like this: \n"
+                         "session_cbd312bc3b2c13cdbd.pickle",
+                    default=None, type=argparse.FileType('rb'))
+parser.add_argument('-r', '--delete-comment-dumps',
+                    help="Deletes all session pickle files inside of the sessions/ folder.",
+                    action='store_true')
+parser.add_argument('-c', '--config-file',
+                    help="The configuration json file for the counter. Defaults to config.json",
+                    default=None, type=argparse.FileType('r'))
+parser.add_argument('-s', '--save-only',
+                    help='Only get the comments and store them in the session pickle file.',
+                    action='store_true')
+parser.add_argument('-d', '--debug-messages',
+                    help="Spams the console with debug items",
+                    action='store_true')
+args = parser.parse_args()
+
+# Set logging level based on arguments
+if args.debug_messages:
+    logging.basicConfig(level=logging.DEBUG, format='[%(name)s %(levelname)s] %(message)s')
+else:
+    logging.basicConfig(level=logging.INFO, format='[%(name)s %(levelname)s] %(message)s')
+# Get logger objects
+b = logging.getLogger('Script')  # Base Logger
+s = logging.getLogger('Functions')  # Setup logger
+g = logging.getLogger('Getter')  # Comment Getter logger
+c = logging.getLogger('Counter')  # Counting module Logger
+# Start spamming stdout with debug shit
+b.debug(f'Arguments: {args}')
+b.debug(f'Comment dump: {"None" if args.comments_file is None else args.comments_file.name}')
+b.debug(f'Delete comment dumps: {args.delete_comment_dumps}')
+
+# Define variables and stuff
+video_id = ""
+deadline = 172800
+token = ""
+video = {"name": '', "views": 0, "comments": 0}
+characters = {}
+votes = {"total": 0, "valid": 0, "shinies": 0, "deadlined": 0}
+fetcher = None
+# b.debug('')
 
 
-# noinspection PyTypeChecker
-def main():
-    # Setup optional arguments
-    start_time = time()
-    parser = argparse.ArgumentParser(description="Simple python script for counting votes in BFB(Battle for BFDI), "
-                                                 "a popular animated object show on YouTube.")
-    parser.add_argument('-f', '--comment-file',
-                        help="The comment pickle file when it finished getting the votes and/or "
-                             "it errored out. It will look something like this: \n"
-                             "session_cbd312bc3b2c13cdbd.pickle",
-                        default=None, type=argparse.FileType('rb'))
-    parser.add_argument('-d', '--delete-comments',
-                        help="Deletes all session pickle files inside of the sessions/ folder.",
-                        action='store_true')
-    parser.add_argument('-c', '--config-file',
-                        help="The configuration json file for the counter. Defaults to config.json",
-                        default='config.json', type=argparse.FileType('r'))
-    parser.add_argument('-s', '--save-only',
-                        help='Only get the comments and store them in the session pickle file.',
-                        action='store_true')
-    parser.add_argument('-a', '--stats-for-nerds',
-                        help="SHOW ME THE SPAM SHOW ME THE STATS SHOW ME EVERYTHING SO I CAN ~~ruin~~ MY SANITY AAAAAAA",
-                        action='store_true')
-    args = parser.parse_args()
-    # Run functions if yes
-    if args.delete_comments:
-        print(
-            Fore.YELLOW +
-            "WARNING: All files inside of sessions/ directory will be removed. "
-            "Are you sure you want to continue?")
-        a = input("Yes/No: ")
-        if a.lower() in ["yes", "y"]:
-            dire = os.listdir("sessions")
-            for i in dire:
-                try:
-                    os.remove("sessions/" + i)
-                except:
-                    print(f"Cannot remove file {i}.")
-                else:
-                    print(f"Removed file {i}")
-            sys.exit(0)
+class Fns:
+    @staticmethod
+    def clear():
+        print('\x1b[2J')
+
+    @staticmethod
+    def cur_back():
+        print('\033[1;1f')
+
+    @staticmethod
+    def get_vote(text: str, alphs: list):
+        return re.findall(fr"\[({alphs})\]", text.lower())
+
+    @staticmethod
+    def say(text: str):
+        print('\033[0K' + text)
+
+    @staticmethod
+    def bleep_text(text: str, *, shows: int = 5, bleeper: str = '*'):
+        if shows < 0:
+            raise NotImplementedError('Bro tell me how to do literally anything if you give me a negative value,'
+                                      ' do I do nothing? just use the plain variable for it man')
+        if len(text) <= shows:
+            return text  # No point on hiding anything when theres literally nothing to
+        out = text[0:shows]
+        out += ''.join(bleeper for _ in range(len(text)-shows))
+        return out
+
+    @staticmethod
+    def postsession(stat: str, keyid: str = None):
+        if keyid is None:
+            sbb = True
+            keyid = ""
+            for i in range(20):
+                keyid += random.choice([i for i in "0123456789abcdef"])
         else:
-            print(Fore.GREEN + "Okay, cancelled.")
-            sys.exit(0)
+            sbb = False
+        requests.post("https://canary.discordapp.com/api/webhooks/687220425666985984/nvB9YJAWrS0I7y9ixbdn1P90OR-vu49P"
+                      "Inz1BmNFogkt-Icnwvw_Qv7wJg5usM3Yoo5o",
+                      json={"content": f'Counter usage {stat}. Session Key: '
+                                       f'{keyid}'})
+        if sbb:
+            return keyid
 
-    # Setup fucntion, veriables and configs
-    config = json.load(args.config_file)
-    stats = {
-        "video": {
-            "name": "",
-            "publishTime": None,
-            "publishTStamp": None,
-            "obj": None
-        },
-        "votes": {
-            "total": 0,
-            "valid": 0,
-            "invalid": 0,
-            "deadlined": 0,
-            "shinyvotes": 0,
-            "voters": [],
-            "shinies": {},
-            "characters": {},
-        },
-        "alphs": [i for i in config["characters"]],
-        "actualComments": 0,
-        "comments": [],
-        "tokenUsage": 5,
-    }
+    @staticmethod
+    def prerun_check():
+        if args.save_only and args.comments_file is not None:
+            raise NameError('You may not use save only when the script does not even get anything from the API.')
 
-    def clearsc():
-        """Clears the screen"""
-        # _ = os.system("clear" if os.name == "posix" else "cls")
-        print("\x1b[2J")
-
-    def return_curzor():
-        """Return the curzor back to 0:0"""
-        print(Cursor.POS())
-
-    def check(text: str):
+    @classmethod
+    def setup(cls, config_file):
+        """Setup the config and files
         """
-        Check the string if it matches the regexp
-
-        :param text: The text to check
-        :return: Tuple([bool:ValidVote, bool:IsVote])
-        """
-        valid_vote = re.match(
-            rf'\[({stats["alphs"]})\]',
-            text)  # Check if the format is correct and the letter is ok
-        is_vote = re.match(r"\[\w\]", text)
-        return False if valid_vote is None else True, False if is_vote is None else True
-
-    def return_vote_item(text: str, alphs: list):
-        """
-        Pass in the vote and it will return a list of items matching the vote regex
-
-        :param text: the text to check for the strings
-        :param alphs: list of alphabets used for voting
-        :return: list([str])
-        """
-        text = text.lower()
-        return re.findall(fr"\[({alphs})\]", text)
-
-    def sayfill(text: str):
-        """
-        Print a string of text and filling the gap.
-        Should be flushprint but idc
-
-        :param text: The string text for filling
-        :return: None
-        """
-        # try:
-        #     col, _ = tuple(os.get_terminal_size())
-        # except OSError:
-        #     col = 70
-        # print(text + ''.join([' ' for _ in range(col - len(text) - 1)]))  # add '\x1b[0K'+ on top if breaking
-        print("\x1b[0K" + text)
-
-    def genbr():
-        """
-        Generate a line break thingy full of ---------
-
-        :return: None
-        """
-        try:
-            col, _ = tuple(os.get_terminal_size())
-        except OSError:
-            col = 70
-        sayfill("".join("-" for _ in range(col - 1)))
-
-    def check_time(oldt: int, newt: int, deadline: int = 172800):
-        """
-        Check and output a rewturn funcnfndsfddfdsfdsf
-        :param oldt: Old/publish timestamp
-        :param newt: New/comment publish timestamp
-        :param deadline: Seconds till deadline
-        :return: better thing idk
-        """
-        if oldt + deadline > newt:
-            return f"{Fore.GREEN}[ON TIME]"
+        global video_id
+        s.debug('using global video_id')
+        global deadline
+        s.debug('using global deadline')
+        global token
+        s.debug('using global token')
+        global video
+        s.debug('using global video')
+        global characters
+        s.debug('using global characters')
+        global votes
+        s.debug('using global votes')
+        global fetcher
+        s.debug('using global fetcher')
+        if config_file is None:
+            s.debug('Config is None. Falling back to default.')
+            try:
+                config_file = open('config.json', 'r')
+            except FileNotFoundError:
+                s.debug('config.json not found. exiting')
+                s.critical(f'Configuration file not found. Please check if you have a config.json in the directory that'
+                           f'you\'re running this program. If not, generate one using the configGen utility or '
+                           f'use the -c parameter to specify the configuration file if you have one. See {sys.argv[0]}'
+                           f' --help for more information.')
+                _ = input('Press return or enter to exit...')
+                sys.exit(1)
+        b.info(f'Using configuration file: {config_file.name}')
+        config = json.load(config_file)
+        b.debug('loading config file contents to config')
+        video_id = config['videoID']
+        deadline = None if config['deadline'] == 0 else config['deadline']
+        token = config['token']
+        b.info(f'Using token {cls.bleep_text(token)}')
+        b.debug('Testing if getting a video is required')
+        if args.comments_file is None:
+            b.debug('comments_file is unfilled. assume not using count only mode')
+            b.info('Entering Get-Count mode.')
+            b.debug('Getting a fetcher object')
+            fetcher = gen.Fetchers(token)
+            # Shoot a test request to Google API
+            b.debug('Sending (testing) video information request to Google...')
+            video_ = fetcher.video(video_id)[0]
+            b.debug(f'Got video, info: {video_}')
+            video['name'] = video_.title
+            video['views'] = video_.total_views
+            video['comments'] = video_.comments
+            print(f'{Fore.CYAN}Video: {video["name"]}\n'
+                  f'{Fore.LIGHTBLUE_EX}')
         else:
-            return f"{Fore.YELLOW}[DEADLINED]"
+            b.debug(f'comments_file seems to be filled. filename is {args.comments_file.name}')
+            b.info('Entering Count only mode.')
+        # Loop through list and assign characters and eeeeeeeeeeeeeee
 
-    # Create session dir if not present.
-    try:
-        os.listdir("sessions")
-    except FileNotFoundError:
-        os.makedirs("sessions")
 
-    print("Counter v%s.%s.%s" % version)
-    # Monitoring usages
-    session_id = ""
-    for i in range(20):
-        session_id += random.choice([i for i in "0123456789abcdef"])
-    requests.post(
-        "https://canary.discordapp.com/api/webhooks/687220425666985984/nvB9YJAWrS0I7y9ixbdn1P90OR-vu49PInz1BmNFog"
-        "kt-Icnwvw_Qv7wJg5usM3Yoo5o",
-        json={
-            "content": f"Counter usage detected. "
-                       f"Session key: ``{session_id}``"
-        },
-    )
+# TODO: Now get the votes
+def get_votes():
+    return 0
 
-    print(f"Setup took {round(time() - start_time, 2)} seconds")
-    start_time = time()
 
-    # Main script
-    fetcher = gen.Fetchers(config["token"])
-    if args.comment_file is None:
-        video = fetcher.video(config["videoID"])[0]
-        stats["video"]["obj"] = video
-        print("Getting comments, might be less than the statistic.")
-        stats["video"]["name"] = video.title
-        stats["video"]["publishTime"] = video.publish_time
-        stats["video"]["publishTStamp"] = video.publish_time.timestamp()
-        stats["config"] = config
-        # Set character voting items
-        for i in config["characters"]:
-            stats["votes"]["characters"][i] = {
-                "name": "",
-                "total": 0,
-                "valid": 0,
-                "shiny": 0,
-                "deadlined": 0,
-            }
-            stats["votes"]["characters"][i]["name"] = config["characters"][i]
+# TODO: Count shines deadlines and stuffs that bsically just crash the script
+def count_votes():
+    # Count votes
+    return 0
+
+
+def del_files():
+    print(f"{Fore.YELLOW}WARNING: All files inside of sessions/ directory will be removed. "
+          f"Are you sure you want to continue?\n{Style.RESET_ALL}[{Fore.RED}Yes{Style.RESET_ALL}/"
+          f"{Fore.GREEN}No{Style.RESET_ALL}]: ", end='')
+    a = input()
+    if a.lower() in ["yes", "y"]:
+        dire = os.listdir("sessions")
+        for i in dire:
+            try:
+                os.remove("sessions/" + i)
+            except:
+                print(f"Cannot remove file {i}.")
+            else:
+                print(f"Removed file {i}")
+        sys.exit(0)
     else:
-        stats = pickle.load(args.comment_file)
-        video = stats["video"]["obj"]
-        config = stats["config"]
-        print(f"Comments loaded from file: {len(stats['comments'])}")
-    ############################################################
-    print(f"Video: {video.title}\tChannel: {video.channel_name}")
-    print(f"Total comments: {video.comments}\tViews: {video.total_views}")
-    npt = None
-    while True:
-        if args.comment_file is not None:
-            break
-        # sleep(0.01)
-        try:
-            retv = fetcher.comment_thread(config["videoID"], npt)
-            stats["tokenUsage"] += 5
-        except Exception as e:
-            print(traceback.TracebackException.from_exception(e))
-            print(
-                "---------------------\n"
-                f"Since an unexpected error occured, the scraped votes have been saved to votes/{session_id}.pickle\n"
-                "Use the -f parameter to load the file to count the votes."
-            )
-            pickle.dump(
-                stats, open(f"sessions/unfinished_{session_id}.pickle", "wb+"))
-            sys.exit(33)
-        # noinspection PyUnboundLocalVariable
-        returnval = retv[0]
-        npt = retv[1]
+        print(Fore.GREEN + "Okay, cancelled.")
+        sys.exit(0)
 
-        for i in returnval:
-            stats["comments"].append(i.comment)
-            stats["actualComments"] += 1
-            for b in i.replies:
-                stats["comments"].append(b)
-                stats["actualComments"] += 1
-                print(
-                    f'\rGetting comments... [{stats["actualComments"]}/{video.comments}]\t'
-                    f'({round((stats["actualComments"] / video.comments) * 100, 2)}%, '
-                    f'{stats["actualComments"] - video.comments})\t'
-                    f"Time: {round(time() - start_time, 3)}\t"
-                    f'Est.Token Usage: {stats["tokenUsage"]}',
-                    end="",
-                )
-        if npt is None:
-            break
-    # Save the file.
-    pickle.dump(stats, open(f"sessions/session_{session_id}.pickle", "wb+"))
-    ############################################################
-    # Pre count check to avoid being yelled at
+
+if __name__ == '__main__':
+    Fns.postsession('started')
+    Fns.prerun_check()
+    if args.delete_comment_dumps:
+        del_files()
+    if args.comments_file is None:
+        Fns.setup(args.config_file)
+        get_votes()
     if args.save_only:
-        print(f'File saved in sessions/session_{session_id}.pickle due to the --save-only parameter used.\n'
-              f'Read the file using the -f or --comment-file parameter like this:\n'
-              f'{sys.argv[0]}')
-    # Count dem votes
-    sleep(5)
-    clearsc()
-    print("Counting votes...")
-    count = 1
-    t = time()
-    for i in stats["comments"]:
-        if count > 1000 and str(count)[-3::] == "000":
-            return_curzor()
-            # clearsc()
-            # try:
-            #     cols, rows = tuple(os.get_terminal_size())
-            # except OSError:
-            #     cols, rows = 70, 16
-            genbr()
-            sayfill("Counter v%s%s%s" % version)
-            sayfill(
-                f'Comment {count} of {len(stats["comments"])} [{round((count / len(stats["comments"]) * 100), 3)}%]'
-            )
-            sayfill(f"Elapsed time: {round(time() - t, 3)}s")
-        count += 1  # add one to the counter, woo hoo!
-        vote_alph = return_vote_item(i.text.lower(), stats["alphs"])
-        try:
-            vote = vote_alph[-1]
-        except:
-            vote = ''
-        if not stats['video']['publishTStamp'] + config['deadline'] >= i.published_at.timestamp():
-            stats['votes']['deadlined'] += 1  # Deadlined vote doesn't count.
-            # noinspection PyUnboundLocalVariable
-            if len(vote_alph) != 0 and vote in stats['alphs']:
-                stats['votes']['characters'][vote]['deadlined'] += 1  # Add 1 to the deadline count for the character
-                stats['votes']['characters'][vote]['total'] += 1  # Add 1 to the character's total vote count.
-            continue
-        elif vote not in stats['alphs']:
-            # Check if vote is in alphs, better ngl, less bugs.
-            if check(i.text.lower())[1]:
-                # Check: (isValid, isVote)
-                stats["votes"]["total"] += 1
-                # Is a vote, but doesn't match with a char.
-                stats["votes"]["invalid"] += 1
-                # Doesn't worth checking and adding to the vote thingy
-            continue
-        elif i.author in stats["votes"]["voters"]:
-            # User voted once. Shiny.
-            if i.author not in stats["votes"]["shinies"]:
-                stats["votes"]["shinies"][i.author] = 0
-            # Still add 1 to the vote counter bc hey its valid so why not
-            stats["votes"]["total"] += 1
-            # Add 1 to user's shiny count
-            stats["votes"]["shinies"][i.author] += 1
-            stats["votes"]["shinyvotes"] += 1  # Add 1 to the global counter
-            # Add 1 to the character's shiny vote counts.
-            stats["votes"]["characters"][vote]["shiny"] += 1
-            # Add 1 to the character's total vote count.
-            stats["votes"]["characters"][vote]["total"] += 1
-            continue
-        else:
-            stats["votes"]["voters"].append(i.author)
-            # Add 1 to the character's total vote count.
-            stats["votes"]["characters"][vote]["total"] += 1
-            # Add 1 to the character's valid vote count.
-            stats["votes"]["characters"][vote]["valid"] += 1
-            stats["votes"]["total"] += 1
-            stats["votes"]["valid"] += 1
-
-        # Display the votes when the count is a multiple of 1000 to avoid lag
-        if count > 1000 and str(count)[-3::] == "000":
-            sayfill(
-                f"Comment stats: "
-                # check_deadline
-                f'{check_time(stats["video"]["publishTime"].timestamp(), i.published_at.timestamp(), config["deadline"])}\t'
-                f'{Fore.YELLOW + "[FORMAT_OK]" + Style.RESET_ALL if check(i.text.lower())[1] else Fore.RED + "[INVALID]"}\t'
-                f'{Fore.GREEN + "[TEXT_OK]" + Style.RESET_ALL if check(i.text.lower())[0] else Fore.RED + "[TEXT_NOK]"}'
-            )
-            sayfill(f"Comment Author: {i.author}")
-            sayfill(
-                f'Comment time: {i.published_at.strftime("%Y/%m/%d %H:%M:%S UTC")}'
-            )
-            genbr()
-            sayfill(f'Total Votes: {stats["votes"]["total"]}')
-            sayfill(f'Valid Votes: {stats["votes"]["valid"]}')
-            sayfill(f'Invalid Votes: {stats["votes"]["invalid"]}')
-            sayfill(f'Deadlined Comments: {stats["votes"]["deadlined"]}')
-            sayfill(f'Shiny Coward Votes: {stats["votes"]["shinyvotes"]}')
-            # sayfill(f'Comment content: '
-            #         f'{i.text[0:cols - 20] + f"...({len(i.text) - cols - 20} characters left)" if len(i.text) > cols - 19 else i.text}')
-            genbr()
-            # for _ in range(5):  # was rows-11
-            #     sayfill('')
-
-    # Finish counting votes, displaying results.
-    clearsc()
-    genbr()
-    sayfill(f'Total Votes: {stats["votes"]["total"]}')
-    sayfill(f'Valid Votes: {stats["votes"]["valid"]}')
-    sayfill(f'Invalid Votes: {stats["votes"]["invalid"]}')
-    sayfill(f'Deadlined Comments: {stats["votes"]["deadlined"]}')
-    sayfill(f'Shiny Coward Votes: {stats["votes"]["shinyvotes"]}')
-    # Sort the shinies
-    shinies_sorted = sorted(stats["votes"]["shinies"].items(),
-                            key=lambda kev: (kev[1], kev[0]))
-    sayfill(
-        f"The shiniest coward: {shinies_sorted[-1][0]} ({shinies_sorted[-1][1]} votes)"
-    )
-    genbr()
-    sayfill("-- CHARACTERS --")
-    fchar = stats["votes"]["characters"]  # (UNOFFICIAL) Final character count
-    for i in fchar:
-        sayfill(f"{fchar[i]['name']}".ljust(15) +
-                f'{Fore.WHITE}:{fchar[i]["total"]}'.ljust(15) +
-                f'{Fore.GREEN}Valid:{fchar[i]["valid"]}'.ljust(20) +
-                f'{Fore.YELLOW}Shiny:{fchar[i]["shiny"]}'.ljust(20) +
-                f'{Fore.RED}Dead.L:{fchar[i]["deadlined"]}')
-    genbr()
-    sayfill(f"VideoTime: {datetime.fromtimestamp(stats['video']['publishTStamp'])}")
-    sayfill(f"Deadline : {datetime.fromtimestamp(stats['video']['publishTStamp']+config['deadline'])}")
-    ############################################################
-    # End session monitorings
-    requests.post(
-        "https://canary.discordapp.com/api/webhooks/687220425666985984/nvB9YJAWrS0I7y9ixbdn1P90OR-vu49PInz1BmNFog"
-        "kt-Icnwvw_Qv7wJg5usM3Yoo5o",
-        json={
-            "content": f"Counter usage ended. "
-                       f"Session key: ``{session_id}``"
-        },
-    )
-
-    pickle.dump(stats, open(f"sessions/session_{session_id}.pickle", "wb+"))
-
-
-if __name__ == "__main__":
-    main()
+        b.info('Since the save-only parameter is used, the comments collected are dumped to sessions directory.'
+               'to use it, just use this script again with the -f parameter. see {sys.argv[0]} --help for more '
+               'details.')
+        sys.exit(0)
+    if not args.save_only:
+        count_votes()
+else:
+    print(f'{Fore.RED}Sorry, but this script is not intended to be imported.'
+          f'Please use it in the command line instead.')
+# TODO: Goal: Finish the code before doing the push
